➕ WOW-44 feat: 검색 헤더에 뒤로가기 추가
