+++ src/app/[locale]/(main)/search/page.tsx
@@ -1,20 +1,7 @@
-import RecentSearchKeyword from "@/components/search/recent-search-keyword";
-import SearchHeader from "@/components/search/search-header";
-import SearchResult from "@/components/search/search-result";
+import SearchContainer from "@/components/search/search-container";
 
 const page = () => {
-  // 예시
-  const keywords = [
-    { id: "1", keyword: "우유" },
-    { id: "2", keyword: "모니터" },
-  ];
-
-  return (
-    <div className="px-4 text-ELSE-33">
-      <SearchHeader />
-      {/* <RecentSearchKeyword keywords={keywords} /> */}
-      <SearchResult />
-    </div>
-  );
+  return <SearchContainer />;
 };
+
 export default page;
